Rename `my_bench` to `forge_bench`
