+++ crates/forge/benches/main.rs
@@ -1,6 +1,6 @@
 use criterion::{criterion_group, criterion_main};
 
-mod my_bench;
+mod forge_bench;
 
-criterion_group!(benches, my_bench::benchmark::criterion_benchmark);
+criterion_group!(benches, forge_bench::benchmark::criterion_benchmark);
 criterion_main!(benches);
